pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discourse-lens"
version = "0.1.0"
description = "LS x ET journal-abstract discourse map (dual-network + topic clustering)"
authors = [{ name = "Jewoong Moon", email = "jewoong.moon@gmail.com" }]
requires-python = ">=3.11"
license = { text = "Proprietary - research use" }
dependencies = [
    "httpx>=0.27",
    "tenacity>=8.2",
    "pydantic>=2.6",
    "typer>=0.12",
    "rich>=13.7",
    "datasketch>=1.6",
    "langdetect>=1.0.9",
    "numpy>=1.26",
    "pandas>=2.2",
]

[project.optional-dependencies]
nlp = [
    "sentence-transformers>=2.7",
    "bertopic>=0.16",
    "umap-learn>=0.5",
    "hdbscan>=0.8",
    "keybert>=0.8",
    "scikit-learn>=1.4",
]
llm = [
    "anthropic>=0.40",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
dlens = "discourse_lens.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
